+++ src/components/admin/SingOutButton.tsx
@@ -3,11 +3,14 @@
 import { logoutAction } from "@/lib/backend/actions/auth-actions";
 import LogoutIcon from "@mui/icons-material/Logout";
 import { Button } from "@mui/material";
+import { useRouter } from "next/navigation";
 
 export default function SignOutButton() {
+  const router = useRouter();
+
   const handleSignOut = async () => {
     await logoutAction();
-    window.location.href = "/";
+    router.push("/");
   };
 
   return (