+++ src/components/region-overview/tract-deed-overview/ResourcePPChart.tsx
@@ -24,7 +24,7 @@ export default function ResourcePPChart({ production }: Props) {
       return indexA - indexB;
     }) as Resource[];
 
-  const rawTraces: Partial< Data>[] = resourceLabels.map((resourceLabel) => {
+  const rawTraces: Partial<Data>[] = resourceLabels.map((resourceLabel) => {
     const rawPP = production[resourceLabel].basePP;
     const color = RESOURCE_COLOR_MAP[resourceLabel] || "black";
     return {