+++ src/lib/backend/actions/region/compare-rarity-actions.ts
@@ -39,8 +39,7 @@ export async function getRegionCompareRarity(
     const grainReqPerHour =
       deed.worksiteDetail?.grain_req_per_hour ?? undefined;
     const resourceRecipe = deed.worksiteDetail?.resource_recipe as unknown as
-      | ResourceRecipeItem[]
-      | [];
+      ResourceRecipeItem[] | [];
 
     const consumeCost = calcCostsWithDEC(
       totalBasePP,